time profiling chnages from Jean-Noel
His explanations for the changes:

In fact I modify the make tools to write the time of each command inside
a file as well as the dependencies.
This works in modt of the time. I think that I forget to print some
information a the beginning of the recurcive call.

I don't have the patch but the code is here :
http://moais.imag.fr/membres/jean-noel.quintin/make-3.82

If you want to do something that is generic and work with every makefile
you have to open a file descriptor (over the file descriptor 3) to write
the profiling.

You could use the code like you want. My goal was to generate some real
workload for the simulation.